+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [2.6.1](https://github.com/bitrix24/b24ui/compare/v2.6.0...v2.6.1) (2026-04-xx)
 
-* docs: improve agent discoverability
+* feat(CommandPalette): add `searchDelay` prop
+
 
 ## [2.6.0](https://github.com/bitrix24/b24ui/compare/v2.5.3...v2.6.0) (2026-04-23)